+++ tests/unit/Filters/GroupTest.php
@@ -145,4 +145,48 @@ class GroupTest extends ModelUnitTestCase
         $this->assertNotEquals("Cuthbert", $model->Surname);
         $this->assertNotEquals("Andrew", $model->Forename);
     }
+
+    public function testAndOrGroupsParametersNotAsArray()
+    {
+        $andGroup = new AndGroup(new Equals("Forename", "John"), new Equals("Surname", "Luc"));
+        $contacts = Example::find($andGroup);
+        $model = $contacts[0];
+        $this->assertEquals("John", $model->Forename);
+        $this->assertEquals("Luc", $model->Surname);
+
+        $orGroup = new OrGroup(new Equals("Forename", "John"), new Equals("Surname", "Luc"));
+        $contacts = Example::find($orGroup);
+        $model = $contacts[0];
+        $this->assertEquals("John", $model->Forename);
+        $this->assertEquals("Joe", $model->Surname);
+
+        $contacts = Example::find(new Equals("Forename", "John"), new Equals("Surname", "Luc"));
+    }
+
+    public function testFindAutomaticAndFilterForSeveralParameters()
+    {
+        $contacts = Example::find(new Equals("Forename", "John"), new Equals("Surname", "Luc"));
+        $model = $contacts[0];
+        $this->assertEquals("John", $model->Forename);
+        $this->assertEquals("Luc", $model->Surname);
+    }
+
+    public function testFilterAutomaticAndFilterForSeveralParameters()
+    {
+        $contacts = $this->list->filter(
+            new Contains("Forename", "Joh", true),
+            new Contains("Surname", "Joh", true)
+        );
+        $model = $contacts[0];
+        $this->assertEquals("John", $model->Forename);
+        $this->assertEquals("Johnson", $model->Surname);
+
+        $contacts2 = $this->list->filter([
+            new Contains("Forename", "Joh", true),
+            new Contains("Surname", "Joh", true)
+        ]);
+        $model = $contacts2[0];
+        $this->assertEquals("John", $model->Forename);
+        $this->assertEquals("Johnson", $model->Surname);
+    }
 }